% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lshtm.R
\name{lshtm}
\alias{lshtm}
\title{LSHTM Beamer Presentation Themes for rmarkdown document}
\usage{
lshtm(
  toc = FALSE,
  slide_level = 2,
  incremental = FALSE,
  fig_width = 8,
  fig_height = 5,
  fig_crop = TRUE,
  fig_caption = TRUE,
  dev = "pdf",
  df_print = "default",
  fonttheme = "default",
  colortheme = "lshtmgreen",
  highlight = "tango",
  keep_tex = FALSE,
  latex_engine = "xelatex",
  citation_package = c("default", "natbib", "biblatex"),
  includes = NULL,
  md_extensions = NULL,
  pandoc_args = NULL
)
}
\arguments{
\item{toc}{A logical variable defaulting to \code{FALSE}.}

\item{slide_level}{A numeric variable defaulting to two.}

\item{incremental}{A logical variable defaulting to \code{FALSE}.}

\item{fig_width}{A numeric variable defaulting to ten.}

\item{fig_height}{A numeric variable defaulting to seven.}

\item{fig_crop}{A logical variable defaulting to \code{TRUE}.}

\item{fig_caption}{A logical variable defaulting to \code{TRUE}.}

\item{dev}{A character variable defaulting to \dQuote{pdf}.}

\item{df_print}{A character variable defaulting to \dQuote{default}.}

\item{fonttheme}{A character variable defaulting to \dQuote{default}.}

\item{colortheme}{A character variable defaulting to \dQuote{lshtmgreen}.
Other possible values: \dQuote{lshtmlightgreen} and \dQuote{lshtmreverse}}

\item{highlight}{A character variable defaulting to \dQuote{tango}.}

\item{keep_tex}{A logical variable defaulting to \code{FALSE}.}

\item{latex_engine}{A character variable defaulting to \dQuote{xelatex}.}

\item{citation_package}{An optional character variable with possible value
\dQuote{default}, \dQuote{natbib} (the default), or \dQuote{biblatex}.}

\item{includes}{An optional character variable defaulting to \code{NULL}.}

\item{md_extensions}{An optional character variable defaulting to \code{NULL}.}

\item{pandoc_args}{An optional character variable defaulting to \code{NULL}.}
}
\value{
RMarkdown content processed is returned for use by the
\code{\link[rmarkdown]{render}} function but the function is invoked
for it side effect of creating the pdf file.
}
\description{
Templates for RMarkdown-based Beamer PDF presentations supporting
the \sQuote{LSHTM} style by David Hajage, entirely based on
the \sQuote{Monash} style by Rob J Hyndman.
}
\details{
The package is based on the 'binb' package by Dirk Eddelbuettel, Ista Zahn and Rob Hyndman.
}
\examples{
\dontrun{
library(rmarkdown)
draft("myslides.Rmd", template="lshtm", package="lshtm", edit=FALSE)
setwd("myslides")       ## template creates a new subdir
render("myslides.Rmd")
}
}
\seealso{
The \code{\link[binb]{binb}} package.
}
\author{
David Hajage
}
